apps/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai>=1.40.0
azure-search-documents==11.4.0
pydantic==2.5.3
python-dotenv==1.0.0

